- Move to sub-folder, and remove django models support. - Add 2 new weather stations. - Update github workflow and README. - Remove Travis CI config file.
