--- .github/workflows/hk0weather-tests.yml
@@ -1,30 +1,0 @@
-name: hk0weather-tests
-on: [push]
-jobs:
-  do-hk0weather-tests:
-    runs-on: ubuntu-latest
-    env:
-      PYTHON: '3.9'
-    steps:
-      - uses: actions/checkout@master
-      - name: Setup python
-        uses: actions/setup-python@master
-        with:
-          python-version: 3.9
-      - name: Install required python packages
-        run: |
-          pip install -r requirements.txt
-      - name: List available scrapers
-        run: |
-          coverage run -m scrapy list
-      - name: Test a scraper of regional weather
-        run: |
-          coverage run -m scrapy crawl regional -o regional.csv:csv
-      - name: Test a scraper of daily weather forecast
-        run: |
-          coverage run -m scrapy crawl hkoforecast -o hkoforecast.csv:csv
-      - name: Test a scraper of 9-day weather forecast
-        run: |
-          coverage run -m scrapy crawl hko9dayforecast -o hko9dayforecast.csv:csv
-      - name: Upload coverage to Codecov
-        uses: codecov/codecov-action@v3
