+++ hk0weather/scrapy.cfg
