+++ requirements.txt
@@ -1,4 +1,2 @@
-Scrapy==2.8.0
-pytz==2019.3
-attrs>=19.2.0
-coverage==5.3
+Scrapy>=2.10.1
+pytz>=2023.3
